Updated Transaction to Transaction Log
-fixed name to be more clear
-Utilizes LinkedList Queue
-Utilizes BufferedReader
-Added TODO comments for late version fixes
